requirements.txt
django
djangorestframework
nltk
numpy
